django/customer/ConsumerProject/requirements.txt
amqp==5.2.0
asgiref==3.7.2
async-timeout==4.0.3
backports.zoneinfo==0.2.1
billiard==4.2.0
celery==5.3.6
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
Django==4.2.9
django-celery-results==2.5.1
djangorestframework==3.14.0
flower==2.0.1
greenlet==3.0.3
humanize==4.9.0
idna==3.6
kombu==5.3.5
prometheus-client==0.19.0
prompt-toolkit==3.0.43
psycopg==3.1.17
python-dateutil==2.8.2
pytz==2023.3.post1
redis==5.0.1
requests==2.31.0
six==1.16.0
SQLAlchemy==2.0.25
sqlparse==0.4.4
tornado==6.4
typing-extensions==4.9.0
tzdata==2023.4
urllib3==2.1.0
vine==5.1.0
wcwidth==0.2.13